+++ eslint.config.cjs
@@ -3,6 +3,9 @@ const eslint = require('@eslint/js');
 const tseslint = require('typescript-eslint');
 
 module.exports = tseslint.config(
+  {
+    ignores: ['out/**', '.vscode-test/**'],
+  },
   eslint.configs.recommended,
   ...tseslint.configs.recommended,
 );